software/app_fiabilidad_ui/requirements.txt
pandas
scipy
pingouin
mermaid-cli
streamlit
matplotlib
seaborn
plotly
numpy
psycopg2-binary